refactor: inject connection token in prisma module
