+++ src/modules/prisma/prisma.config.ts
@@ -12,3 +12,5 @@ export type PrismaModuleOptions<
 };
 
 export const { PRISMA_CLIENT_TOKEN } = defineToken('PRISMA_CLIENT_TOKEN');
+
+export const { MONGO_CONNECTION_TOKEN } = defineToken('MONGO_CONNECTION_TOKEN');
